Loan Decision Code 2

+++ src/main/java/ee/forInbank/homeWork/loanDecisionController/LoanDecisionController.java
@@ -14,11 +14,15 @@ public class LoanDecisionController {
     @Autowired
     private LoanDecisionService loanDecisionService;
 
-//    http://localhost:8080/loandecision?customerId=49002010998&loanAmount=4000&loanPeriod=12
+//    TODO testcase 1: http://localhost:8080/loandecision?customerid=49002010965&loanamount=6000&loanperiod=12
+//    TODO testcase 2: http://localhost:8080/loandecision?customerid=49002010976&loanamount=6000&loanperiod=12
+//    TODO testcase 3: http://localhost:8080/loandecision?customerid=49002010987&loanamount=6000&loanperiod=12
+//    TODO testcase 4: http://localhost:8080/loandecision?customerid=49002010998&loanamount=6000&loanperiod=12
+
     @GetMapping("loandecision")
     public String loanSum(@RequestParam("customerid") Double customerId,
                           @RequestParam("loanamount") Double loanAmount,
-                          @RequestParam("loanperiod") Double loanPeriod) {
+                          @RequestParam("loanperiod") Integer loanPeriod) {
         return loanDecisionService.loanSum(customerId, loanAmount, loanPeriod);
     }
 }
